Move signIn/register dialogs to `common/` folder
