+++ packages/website/src/common/RegisterDialog/index.test.tsx
@@ -3,7 +3,7 @@ import { render } from "@testing-library/react";
 import { Provider } from "react-redux";
 import configureStore from "redux-mock-store";
 import { BrowserRouter as Router } from "react-router-dom";
-import { mockUser } from "../../../mocks/mockUser";
+import { mockUser } from "../../mocks/mockUser";
 
 import RegisterDialog from ".";
