+++ packages/website/src/common/SignInDialog/index.tsx
@@ -31,9 +31,9 @@ import {
   userInfoSelector,
   userLoadingSelector,
   userErrorSelector,
-} from "../../../store/User/userSlice";
-import { UserSignInParams } from "../../../store/User/types";
-import incomingStyles from "../styles";
+} from "../../store/User/userSlice";
+import { UserSignInParams } from "../../store/User/types";
+import incomingStyles from "../../routes/Homepage/styles";
 
 const SignInDialog = ({
   open,
